requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
openpyxl
plotly
firebase-admin 
pyrebase4
setuptools
